Move Jira service to new jbi.jira.service module
