+++ jbi/jira/__init__.py
@@ -1,0 +1,3 @@
+from .service import JiraService, get_service
+
+__all__ = ["JiraService", "get_service"]
