--- jbi/services/__init__.py
@@ -1,5 +1,0 @@
-"""Exports the bugzilla and jira services modules for convenience"""
-
-from . import jira
-
-__all__ = ["jira"]
