+++ jbi/steps.py
@@ -35,8 +35,8 @@ class StepStatus(Enum):
 # https://docs.python.org/3.11/library/typing.html#typing.TYPE_CHECKING
 if TYPE_CHECKING:
     from jbi.bugzilla.service import BugzillaService
+    from jbi.jira.service import JiraService
     from jbi.models import ActionContext, ActionParams
-    from jbi.services.jira import JiraService
 
     StepResult = tuple[StepStatus, ActionContext]
