+++ tests/conftest.py
@@ -10,12 +10,11 @@ from fastapi.testclient import TestClient
 from pytest_factoryboy import register
 
 import tests.fixtures.factories as factories
-from jbi import Operation, bugzilla
+from jbi import Operation, bugzilla, jira
 from jbi.app import app
 from jbi.configuration import get_actions
 from jbi.environment import Settings
 from jbi.models import ActionContext
-from jbi.services import jira
 
 
 class FilteredLogCaptureFixture(pytest.LogCaptureFixture):
@@ -109,7 +108,7 @@ def mocked_jira(request):
         yield None
         jira.get_service.cache_clear()
     else:
-        with mock.patch("jbi.services.jira.JiraClient") as mocked_jira:
+        with mock.patch("jbi.jira.service.JiraClient") as mocked_jira:
             yield mocked_jira()
             jira.get_service.cache_clear()
